requirements.txt
Pillow==10.1.0
psutil==5.9.4
requests==2.28.2
urllib3==1.26.14
